requirements.txt
PyJWT==2.3.0
pytest==6.2.5
requests==2.26.0
